Narrow double-to-int conversion at boundary and strip redundant operand promotions (F11/tarefa 11)
In C++, narrowing conversions (such as double -> int) happen at the boundary
(variable initialization, assignment, return, argument passing), while usual
arithmetic promotions inside binary expressions (e.g. `int * double` -> `double`)
are evaluated entirely in double.

Previously, lower_expr lowered each implicit IntegralToFloating cast into an
Expr::Convert{ ty: Double } on every binary operand, and emit_convert_operand
did not parenthesize binary operands when suffixing `.toInt()`. Because member
access binds tighter than binary arithmetic in Dart, `int x = a * 0.5 + b;` was
emitted unparenthesized as `int x = a.toDouble() * 0.5 + b.toDouble().toInt();`,
chaining `.toInt()` onto the rightmost operand as a no-op while leaving the
overall expression typed as `double`.

- lower_binary_operand: strips implicit Int -> Double promotions on operands
  of binary arithmetic and comparison operators, preserving explicit casts
  (such as static_cast<double>).
- lower_expr: cancels chained int -> double -> int conversions during IR
  construction.
- emit_convert_operand: wraps operands in parentheses if their outermost operator
  has lower precedence than postfix call / member access.
- emit_expr: prioritizes unwrapping T? -> T nullable scalar targets with !
  before falling back to scalar numeric conversions.
- ir::Expr: adds Expr::ty(&self) for unified expression type retrieval.

+++ examples/E13-fatia-real-verovio/expected/lib/fraction.dart
@@ -53,7 +53,7 @@ class Fraction {
   }
 
   double ToDouble() {
-    return _m_numerator.toDouble() / _m_denominator.toDouble();
+    return _m_numerator.toDouble() / _m_denominator;
   }
 
   void Reduce() {
